APR Updates
Two key changes here: 1) Include a separate function to call the users VeBAL balance for voting conditions. This is always less than or equal to their veBAL balance for boost due to delegations. 2) APR is calculated within the BribeSimulation index onChange when a pool is selected. This relies on a useEffect hook for updating.