Update Contracts on Map
Includes additional contracts on the map itself. Still needs arrows for permissions and interactions updated. Can create a legend for different line types. Feel free to double check, some things such as gauge and pool factories are not so granular, maybe we can put sub classes for each type. (weighted, composable, etc.) If not no harm as the list is clear below.